+++ CHANGELOG.md
@@ -1,7 +1,17 @@
+# 2.0.0
+
+* Updated apache2 cookbook from 3.1.0 to 3.3.0
+* Updated chef-vault cookbook from 1.3.0 to 2.1.1
+* Updated database cookbook from 4.0.9 to 6.1.1
+* Updated postgresql cookbook from 4.0 to 6.1.1
+* Updated tar cookbook from 0.7.0 to 2.0.0
+* Updated php cookbook from 1.7.2 to 4.0.0
+* Cleaned up some lint issues
+
 # 1.4.1
 
 * Fixed test-kitchen not working
-* Added checksum to mediawiki download to stop it from downloading everytime Chef runs
+* Added checksum to mediawiki download to stop it from downloading every time Chef runs
 
 # 1.0.0